+++ ambry-api/src/main/java/com.github.ambry/network/RequestHandler.java
@@ -11,11 +11,10 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.server;
+package com.github.ambry.network;
 
-import com.github.ambry.network.Request;
-import com.github.ambry.network.RequestResponseChannel;
-import com.github.ambry.utils.Utils;
+import com.github.ambry.server.EmptyRequest;
+import com.github.ambry.server.RequestAPI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -26,10 +25,10 @@ import org.slf4j.LoggerFactory;
 public class RequestHandler implements Runnable {
   private final int id;
   private final RequestResponseChannel requestChannel;
-  private final AmbryRequests requests;
+  private final RequestAPI requests;
   private Logger logger = LoggerFactory.getLogger(getClass());
 
-  public RequestHandler(int id, RequestResponseChannel requestChannel, AmbryRequests requests) {
+  public RequestHandler(int id, RequestResponseChannel requestChannel, RequestAPI requests) {
     this.id = id;
     this.requestChannel = requestChannel;
     this.requests = requests;
@@ -59,35 +58,3 @@ public class RequestHandler implements Runnable {
   }
 }
 
-// Request handler pool. A pool of threads that handle requests
-class RequestHandlerPool {
-
-  private Thread[] threads = null;
-  private RequestHandler[] handlers = null;
-  private Logger logger = LoggerFactory.getLogger(getClass());
-
-  public RequestHandlerPool(int numThreads, RequestResponseChannel requestResponseChannel, AmbryRequests requests) {
-    threads = new Thread[numThreads];
-    handlers = new RequestHandler[numThreads];
-    for (int i = 0; i < numThreads; i++) {
-      handlers[i] = new RequestHandler(i, requestResponseChannel, requests);
-      threads[i] = Utils.daemonThread("request-handler-" + i, handlers[i]);
-      threads[i].start();
-    }
-  }
-
-  public void shutdown() {
-    try {
-      logger.info("shutting down");
-      for (RequestHandler handler : handlers) {
-        handler.shutdown();
-      }
-      for (Thread thread : threads) {
-        thread.join();
-      }
-      logger.info("shut down completely");
-    } catch (Exception e) {
-      logger.error("error when shutting down request handler pool {}", e);
-    }
-  }
-}
