+++ ambry-api/src/main/java/com.github.ambry/server/RequestAPI.java
@@ -62,4 +62,11 @@ public interface RequestAPI {
    * @throws InterruptedException
    */
   void handleTtlUpdateRequest(Request request) throws IOException, InterruptedException;
+
+  /**
+   * Handles the given request.
+   * @param request {@code Request} to be handled.
+   * @throws InterruptedException
+   */
+  public void handleRequests(Request request) throws InterruptedException;
 }
