+++ ambry-commons/src/main/java/com.github.ambry.commons/ErrorMapping.java
@@ -11,9 +11,8 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.server;
+package com.github.ambry.commons;
 
-import com.github.ambry.commons.ServerErrorCode;
 import com.github.ambry.messageformat.MessageFormatErrorCodes;
 import com.github.ambry.store.StoreErrorCodes;
 import java.util.Collections;
